Remove CodeQL analysis configuration
The `codeql.yml` file has been removed from the repository. This file was responsible for setting up and running CodeQL analysis on the codebase, including specifying the languages to analyze, the schedule for running the analysis, and the steps to perform the analysis.

Additionally, the reference to `.github\workflows\codeql.yml` has been removed from the `SolutionItems` section in the `Mtd.Stopwatch.Infrastructure.EFCore.sln` project file. This indicates that the CodeQL workflow is no longer part of the solution's workflow configuration.